[ui-components] refactor: added checkbox/deopdown/infoicon
