+++ packages/editor-common/web/package.json
@@ -40,7 +40,9 @@
     "downshift": "3.2.12",
     "focus-trap": "6.4.0",
     "react-window": "~1.5.0",
-    "wix-rich-content-common": "8.44.1"
+    "wix-rich-content-common": "8.44.1",
+    "wix-rich-content-ui-components": "8.44.1"
+    
   },
   "unpkg": true,
   "publishConfig": {
