+++ packages/plugin-html/web/src/AdsenseTitle.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import styles from '../statics/styles/adsense.scss';
-import { InfoIcon } from 'wix-rich-content-editor-common';
+import { InfoIcon } from 'wix-rich-content-ui-components';
 import { TranslationFunction } from 'wix-rich-content-common';
 
 const AdsenseTitle: FunctionComponent<{ t: TranslationFunction }> = ({ t }) => {
