+++ packages/plugin-social-polls/web/src/components/settings/layout-settings-section/LayoutSettingsSection.jsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import { LTRIcon, RTLIcon } from 'wix-rich-content-plugin-commons';
-import { SelectionList, LabeledToggle } from 'wix-rich-content-ui-components';
-import { Separator, InfoIcon } from 'wix-rich-content-editor-common';
+import { SelectionList, LabeledToggle, InfoIcon } from 'wix-rich-content-ui-components';
+import { Separator } from 'wix-rich-content-editor-common';
 import { mergeStyles } from 'wix-rich-content-common';
 
 import { LAYOUT, DIRECTION } from '../../../defaults';
