+++ packages/toolbars/web/package.json
@@ -36,7 +36,8 @@
   "dependencies": {
     "wix-rich-content-common": "8.44.1",
     "wix-rich-content-editor-common": "8.44.1",
-    "wix-rich-content-plugin-commons": "8.44.1"
+    "wix-rich-content-plugin-commons": "8.44.1",
+    "wix-rich-content-ui-components": "8.44.1"
   },
   "unpkg": true,
   "publishConfig": {
