+++ packages/ui-components/web/src/Components/InputWithLabel.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { omit } from 'lodash';
 import classNames from 'classnames';
-import { InfoIcon } from 'wix-rich-content-editor-common';
+import InfoIcon from './InfoIcon';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/input-with-label.scss';
 import generalstyles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
