+++ packages/ui-components/web/src/Components/RadioGroupHorizontal.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { RadioGroup, InfoIcon } from 'wix-rich-content-editor-common';
+import { RadioGroup } from 'wix-rich-content-editor-common';
+import InfoIcon from './InfoIcon';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/radio-group-horizontal.scss';
 import generalstyles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
