+++ packages/ui-components/web/src/Icons/index.ts
@@ -1,0 +1,3 @@
+export { default as DropdownArrowIcon } from './DropdownArrowIcon';
+export { default as CheckIcon } from './CheckIcon';
+export { default as InfoIcon } from './InfoIcon';
