+++ packages/ui-components/web/statics/styles/checkbox.scss
