+++ packages/ui-components/web/statics/styles/dropdown.scss
